add common appbar and update bottom navigation profile tab
