+++ lib/core/router/app_shell.dart
@@ -86,11 +86,20 @@ class AppShell extends StatelessWidget {
                 label: "내 캘린더",
               ),
 
-              // TODO: 프로필 미 연결 추후 연결 필요
               // 프로필 - active: false(클릭되지 않은) / true(클릭된)
               BottomNavigationBarItem(
-                icon: Icon(Icons.person_outline, color: Colors.grey),
-                activeIcon: Icon(Icons.person, color: Colors.blue),
+                icon: _buildSharedCalendarIcon(
+                  active: false,
+                  icon: Icon(
+                    Icons.person_outline,
+                    size: 22,
+                    color: Colors.white,
+                  ),
+                ),
+                activeIcon: _buildSharedCalendarIcon(
+                  active: true,
+                  icon: Icon(Icons.person, size: 22, color: Colors.white),
+                ),
                 label: "프로필",
               ),
             ],
