+++ lib/core/widgets/back_actions_app_bar.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+
+class BackActionsAppBar extends StatelessWidget implements PreferredSizeWidget {
+  final Widget title;
+  final VoidCallback? onAction;
+  final List<Widget>? actions;
+
+  const BackActionsAppBar({
+    super.key,
+    required this.title,
+    this.onAction,
+    this.actions,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        border: Border(
+          bottom: BorderSide(color: Colors.grey.shade400, width: 1),
+        ),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.08),
+            blurRadius: 4.0,
+            offset: const Offset(0, 1),
+          ),
+        ],
+      ),
+      child: AppBar(
+        backgroundColor: Colors.white,
+        leading: const BackButton(),
+        title: title,
+        titleSpacing: 0,
+        actions: actions,
+      ),
+    );
+  }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(kToolbarHeight);
+}
