+++ lib/features/calendar/presentation/views/shared/shared_calendar_list_screen.dart
@@ -1,32 +1,37 @@
-import 'dart:ui';
-
+import 'package:dutytable/features/calendar/presentation/viewmodels/shared_calendar_view_model.dart';
 import 'package:flutter/material.dart';
-import '../../widgets/duty_table_appbar.dart';
+import 'package:go_router/go_router.dart';
+import 'package:provider/provider.dart';
 
-class SharedCalendarListScreen extends StatefulWidget {
+import '../../../../../core/widgets/logo_actions_app_bar.dart';
+import '../../widgets/calendar_card.dart';
+
+/// TODO
+/// 캘린더 목록 API 연결 시 하드코딩 삭제 필요
+class SharedCalendarListScreen extends StatelessWidget {
   const SharedCalendarListScreen({super.key});
 
   @override
-  State<SharedCalendarListScreen> createState() =>
-      _SharedCalendarListScreenState();
+  Widget build(BuildContext context) {
+    return ChangeNotifierProvider(
+      create: (_) => SharedCalendarViewModel(),
+      child: _SharedCalendarListScreen(),
+    );
+  }
 }
 
-class _SharedCalendarListScreenState extends State<SharedCalendarListScreen> {
-  bool deleteMode = false; // 삭제 모드 ON/OFF
-  bool isSelected = false; // 카드 체크 상태
-  bool isAdmin = false;
+class _SharedCalendarListScreen extends StatelessWidget {
+  const _SharedCalendarListScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
+    final viewModel = context.watch<SharedCalendarViewModel>();
+
     return Scaffold(
       backgroundColor: Colors.white,
-      appBar: DutyTableAppBar(
-        deleteMode: deleteMode,
-        onDeletePressed: () {
-          setState(() {
-            deleteMode = !deleteMode;
-          });
-        },
+      appBar: LogoActionsAppBar(
+        leftActions: _LeftActions(),
+        rightActions: _RightActions(),
       ),
       body: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 20.0),
@@ -34,178 +39,133 @@ class _SharedCalendarListScreenState extends State<SharedCalendarListScreen> {
           children: [
             const SizedBox(height: 10),
 
-            _buildCalendarCard(
+            CalendarCard(
               title: "팀 프로젝트",
-              deletable: true,
-              isAdmin: false,
+              deleteMode: viewModel.deleteMode,
+              isAdmin: viewModel.isAdmin,
+              isSelected: viewModel.isSelected("team_project_1"),
+              onChangeSelected: () =>
+                  viewModel.toggleSelected("team_project_1"),
             ),
 
             const SizedBox(height: 12),
 
-            _buildCalendarCard(title: "부서 공지", deletable: false, isAdmin: true),
+            CalendarCard(
+              title: "부서 공지",
+              deleteMode: viewModel.deleteMode,
+              isAdmin: !viewModel.isAdmin,
+              isSelected: viewModel.isSelected("dept_notice"),
+              onChangeSelected: () => viewModel.toggleSelected("dept_notice"),
+            ),
+
+            const SizedBox(height: 12),
+
+            CalendarCard(
+              title: "운동",
+              deleteMode: viewModel.deleteMode,
+              isAdmin: viewModel.isAdmin,
+              isSelected: viewModel.isSelected("exercise"),
+              onChangeSelected: () => viewModel.toggleSelected("exercise"),
+            ),
           ],
         ),
       ),
     );
   }
+}
 
-  Widget _buildCalendarCard({
-    required String title,
-    required bool deletable,
-    required bool isAdmin,
-  }) {
-    final applyBlur = deleteMode && isAdmin;
+/// 앱바 - 왼쪽 로고 + 타이틀
+class _LeftActions extends StatelessWidget {
+  const _LeftActions({super.key});
 
-    return Stack(
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisSize: MainAxisSize.min,
       children: [
-        Card(
-          elevation: 2,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(12),
-            side: BorderSide(
-              color: isSelected && !applyBlur ? Colors.red : Colors.grey,
-              width: isSelected && !applyBlur ? 2 : 1,
-            ),
-          ),
-          child: Container(
-            padding: const EdgeInsets.all(16),
-            width: double.infinity,
-            child: Row(
-              children: [
-                // LEFT ICON
-                Container(
-                  width: 40,
-                  height: 40,
-                  decoration: BoxDecoration(
-                    color: const Color(0xFF3A7BFF),
-                    borderRadius: BorderRadius.circular(16),
-                  ),
-                  child: Image.asset(
-                    "assets/images/calendar_logo.png",
-                    fit: BoxFit.contain,
-                  ),
-                ),
-
-                const SizedBox(width: 16),
-
-                // MIDDLE TITLE & DATE
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      // 제목 + 멤버수 뱃지
-                      Row(
-                        children: [
-                          Text(
-                            "팀 프로젝트",
-                            style: const TextStyle(
-                              fontSize: 14,
-                              fontWeight: FontWeight.w700,
-                            ),
-                          ),
-                          const SizedBox(width: 8),
-                          Container(
-                            padding: const EdgeInsets.symmetric(
-                              horizontal: 12,
-                              vertical: 2,
-                            ),
-                            decoration: BoxDecoration(
-                              color: Colors.grey.shade200,
-                              borderRadius: BorderRadius.circular(16),
-                            ),
-                            child: Text(
-                              "99+명",
-                              style: const TextStyle(
-                                fontSize: 10,
-                                fontWeight: FontWeight.w600,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-
-                      const SizedBox(height: 6),
-
-                      // 다음 일정
-                      Text(
-                        "다음 일정: 12월 5일",
-                        style: TextStyle(
-                          fontSize: 12,
-                          color: Colors.grey.shade600,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-                const SizedBox(width: 16),
-
-                // RIGHT SIDE
-                if (!deleteMode)
-                  // 기본 뱃지
-                  Container(
-                    width: 42,
-                    height: 42,
-                    decoration: const BoxDecoration(
-                      color: Colors.red,
-                      shape: BoxShape.circle,
-                    ),
-                    alignment: Alignment.center,
-                    child: const Text(
-                      "99+",
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  )
-                else
-                  // 삭제 모드일 때
-                  !isAdmin
-                      ? Checkbox(
-                          value: isSelected,
-                          activeColor: Colors.red,
-                          onChanged: (value) {
-                            setState(() {
-                              isSelected = value ?? false;
-                            });
-                          },
-                        )
-                      : Container(
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 10,
-                            vertical: 4,
-                          ),
-                          decoration: BoxDecoration(
-                            color: Colors.grey.shade300,
-                            borderRadius: BorderRadius.circular(16),
-                          ),
-                          child: const Text(
-                            "삭제 불가",
-                            style: TextStyle(
-                              fontSize: 12,
-                              fontWeight: FontWeight.w600,
-                              color: Colors.black54,
-                            ),
-                          ),
-                        ),
-              ],
-            ),
-          ),
+        Image.asset(
+          "assets/images/calendar_logo.png",
+          width: 60,
+          height: 60,
+          fit: BoxFit.fill,
         ),
-
-        if (applyBlur)
-          Positioned.fill(
-            child: IgnorePointer(
-              child: Container(
-                padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(16),
-                  color: Colors.white.withOpacity(0.6), // 흐릿한 흰색 오버레이
-                ),
-              ),
-            ),
-          ),
+        const SizedBox(width: 4),
+        const Text(
+          "DutyTable",
+          style: TextStyle(fontSize: 20, fontWeight: FontWeight.w700),
+        ),
       ],
     );
+  }
+}
+
+/// 앱바 - 오른쪽 아이콘 모음들
+class _RightActions extends StatelessWidget {
+  const _RightActions({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final viewModel = context.watch<SharedCalendarViewModel>();
+
+    return viewModel.deleteMode
+        ? Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              GestureDetector(
+                onTap: () => viewModel.cancelDeleteMode(),
+                child: Text("취소"),
+              ),
+              const SizedBox(width: 16),
+              Text("삭제"),
+            ],
+          )
+        : Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Container(
+                width: 42,
+                height: 42,
+                decoration: const BoxDecoration(
+                  color: Color(0xFF3A7BFF),
+                  shape: BoxShape.circle,
+                ),
+                child: GestureDetector(
+                  onTap: () => context.push("/shared/add"),
+                  child: Icon(Icons.add, color: Colors.white),
+                ),
+              ),
+
+              const SizedBox(width: 16),
+
+              GestureDetector(
+                onTap: () => context.push("/notification"),
+                child: Stack(
+                  clipBehavior: Clip.none,
+                  children: [
+                    const Icon(Icons.notifications_none, size: 26),
+                    Positioned(
+                      right: -1,
+                      top: -1,
+                      child: Container(
+                        width: 10,
+                        height: 10,
+                        decoration: const BoxDecoration(
+                          color: Colors.red,
+                          shape: BoxShape.circle,
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+
+              const SizedBox(width: 16),
+
+              GestureDetector(
+                onTap: () => viewModel.toggleDeleteMode(),
+                child: const Icon(Icons.delete_outline, size: 26),
+              ),
+            ],
+          );
   }
 }
