+++ lib/features/notification/presentation/widgets/all_delete_dialog.dart
@@ -1,0 +1,90 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+class AllDeleteDialog extends StatelessWidget {
+  /// 알림 전체 삭제 다이얼로그
+  const AllDeleteDialog({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final size = MediaQuery.of(context).size;
+
+    return Center(
+      child: Material(
+        color: Colors.transparent,
+        child: Container(
+          width: size.width * 0.85,
+          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 20),
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(16),
+            boxShadow: [
+              BoxShadow(blurRadius: 20, color: Colors.black.withOpacity(0.15)),
+            ],
+          ),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                "전체 알림을 삭제하시겠습니까?",
+                style: TextStyle(fontSize: 18, fontWeight: FontWeight.w800),
+              ),
+
+              const SizedBox(height: 14),
+
+              Row(
+                children: [
+                  Expanded(
+                    child: GestureDetector(
+                      onTap: () => context.pop(),
+                      child: Container(
+                        decoration: BoxDecoration(
+                          color: Color(0xfff3f4f6),
+                          borderRadius: BorderRadius.circular(10),
+                        ),
+                        alignment: Alignment.center,
+                        padding: EdgeInsets.symmetric(vertical: 10),
+                        child: const Text(
+                          "취소",
+                          style: TextStyle(
+                            fontSize: 14,
+                            fontWeight: FontWeight.w600,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+
+                  const SizedBox(width: 10),
+
+                  Expanded(
+                    child: GestureDetector(
+                      /// TODO: 알림 전체(초대 및 리마인더) 삭제 API 연동 필요
+                      onTap: () {},
+                      child: Container(
+                        decoration: BoxDecoration(
+                          color: Color(0xffef4444),
+                          borderRadius: BorderRadius.circular(10),
+                        ),
+                        alignment: Alignment.center,
+                        padding: EdgeInsets.symmetric(vertical: 10),
+                        child: const Text(
+                          "확인",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 14,
+                            fontWeight: FontWeight.w600,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
