+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   supabase_flutter: ^2.10.3
   intl: ^0.20.2
   url_launcher: ^6.3.2
+  image_picker: ^1.1.0
 
 
 dev_dependencies:
